Move TestSubqueryPerformanceOnLazyExecutionMode to db package and disable it on Travis
